refactor(presenter): split ratatui presenter module
